Added additional explanations | Sections 2.3 and 2.5 of Phase 2 are now bonus
Added video tutorials on API and LLM foundational concepts to understand Phase 2 content better. Now sections 2.3 Information Extraction and Grounding and 2.5 Retrieval-Augmented Generation (RAG) will be considered bonus sections, and as such their exercises will not be counted for the main graded, they will be extra points counted together with the exercises in the Bonus notebook. Fixing smaller errors and revising the spec description in the homework notebook for easier understanding.